# Pipe auditing mixin now ensures named pipes start with a backslash