The `named_pipes.txt` wordlist provided by the metasploit framework contains pipe names that are not prefixed with a backslash. The `auxiliary/scanner/smb/pipe_auditor` module does not seem to work properly for some Samba protocol versions (namely protocol 3) unless each pipe name is prefixed with a backslash. I updated `lib/exploit/smb/client/pipe_auditor.rb` to check for the missing backslash and to prepend it if missing.

## Verifications

List the steps needed to make sure this thing works

- [x] Start `msfconsole`
- [x] `use auxiliary/scanner/smb/pipe_auditor`
- [x] **Verify** the thing does what it should
     Running with default named_pipes.txt does not work without fix and does with fix
- [x] **Document** (nothing to document or no existing documents to add to)

Fixes #9618 and fixes #11870.