@wvu-r7 2019-05-21
Can you post your output and Windows version?

It is working for me on `master` against Windows Server 2008 R2:

```
msf5 auxiliary(scanner/smb/pipe_auditor) > git status
[*] exec: git status

On branch master
Your branch is up to date with 'upstream/master'.

nothing to commit, working tree clean
msf5 auxiliary(scanner/smb/pipe_auditor) > options

Module options (auxiliary/scanner/smb/pipe_auditor):

   Name         Current Setting                                              Required  Description
   ----         ---------------                                              --------  -----------
   NAMED_PIPES  /rapid7/metasploit-framework/data/wordlists/named_pipes.txt  yes       List of named pipes to check
   RHOSTS       192.168.56.103                                               yes       The target address range or CIDR identifier
   SMBDomain    .                                                            no        The Windows domain to use for authentication
   SMBPass      trustno1!                                                    no        The password for the specified username
   SMBUser      Administrator                                                no        The username to authenticate as
   THREADS      1                                                            yes       The number of concurrent threads

msf5 auxiliary(scanner/smb/pipe_auditor) > cat /rapid7/metasploit-framework/data/wordlists/named_pipes.txt
[*] exec: cat /rapid7/metasploit-framework/data/wordlists/named_pipes.txt

netlogon
lsarpc
samr
browser
atsvc
DAV RPC SERVICE
epmapper
eventlog
InitShutdown
keysvc
lsass
LSM_API_service
ntsvcs
plugplay
protected_storage
router
SapiServerPipeS-1-5-5-0-70123
scerpc
srvsvc
tapsrv
trkwks
W32TIME_ALT
wkssvc
PIPE_EVENTROOT\CIMV2SCM EVENT PROVIDER
db2remotecmd
msf5 auxiliary(scanner/smb/pipe_auditor) > run

[*] 192.168.56.103:445    - Connected to named pipe: netlogon
[*] 192.168.56.103:445    - Connected to named pipe: lsarpc
[*] 192.168.56.103:445    - Connected to named pipe: samr
[-] 192.168.56.103:445    - Inaccessible named pipe: browser - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 192.168.56.103:445    - Connected to named pipe: atsvc
[-] 192.168.56.103:445    - Inaccessible named pipe: DAV RPC SERVICE - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 192.168.56.103:445    - Connected to named pipe: epmapper
[*] 192.168.56.103:445    - Connected to named pipe: eventlog
[*] 192.168.56.103:445    - Connected to named pipe: InitShutdown
[*] 192.168.56.103:445    - Connected to named pipe: keysvc
[*] 192.168.56.103:445    - Connected to named pipe: lsass
[*] 192.168.56.103:445    - Connected to named pipe: LSM_API_service
[*] 192.168.56.103:445    - Connected to named pipe: ntsvcs
[*] 192.168.56.103:445    - Connected to named pipe: plugplay
[*] 192.168.56.103:445    - Connected to named pipe: protected_storage
[-] 192.168.56.103:445    - Inaccessible named pipe: router - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 192.168.56.103:445    - Inaccessible named pipe: SapiServerPipeS-1-5-5-0-70123 - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 192.168.56.103:445    - Connected to named pipe: scerpc
[*] 192.168.56.103:445    - Connected to named pipe: srvsvc
[-] 192.168.56.103:445    - Inaccessible named pipe: tapsrv - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 192.168.56.103:445    - Connected to named pipe: trkwks
[*] 192.168.56.103:445    - Connected to named pipe: W32TIME_ALT
[*] 192.168.56.103:445    - Connected to named pipe: wkssvc
[-] 192.168.56.103:445    - Inaccessible named pipe: PIPE_EVENTROOT\CIMV2SCM EVENT PROVIDER - The server responded with error: STATUS_ACCESS_DENIED (Command=162 WordCount=0)
[-] 192.168.56.103:445    - Inaccessible named pipe: db2remotecmd - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[+] 192.168.56.103:445    - Pipes: \netlogon, \lsarpc, \samr, \atsvc, \epmapper, \eventlog, \InitShutdown, \keysvc, \lsass, \LSM_API_service, \ntsvcs, \plugplay, \protected_storage, \scerpc, \srvsvc, \trkwks, \W32TIME_ALT, \wkssvc
[*] 192.168.56.103:       - Scanned 1 of 1 hosts (100% complete)
[*] Auxiliary module execution completed
msf5 auxiliary(scanner/smb/pipe_auditor) >
```

Are you sure the target has SMBv1 enabled? I don't think this code was updated to use RubySMB yet, so SMBv2 does not work.

```
msf5 auxiliary(scanner/smb/pipe_auditor) > set rhosts 192.168.56.101
rhosts => 192.168.56.101
msf5 auxiliary(scanner/smb/pipe_auditor) > run

[*] 192.168.56.101:       - Scanned 1 of 1 hosts (100% complete)
[*] Auxiliary module execution completed
msf5 auxiliary(scanner/smb/pipe_auditor) >
```

```
wvu@kharak:~$ nmap -Pn -n -v -p 445 --script smb-protocols 192.168.56.101,103
Starting Nmap 7.70 ( https://nmap.org ) at 2019-05-21 14:30 CDT
NSE: Loaded 1 scripts for scanning.
NSE: Script Pre-scanning.
Initiating NSE at 14:30
Completed NSE at 14:30, 0.00s elapsed
Initiating Connect Scan at 14:30
Scanning 2 hosts [1 port/host]
Discovered open port 445/tcp on 192.168.56.103
Discovered open port 445/tcp on 192.168.56.101
Completed Connect Scan at 14:30, 0.00s elapsed (2 total ports)
NSE: Script scanning 2 hosts.
Initiating NSE at 14:30
Completed NSE at 14:30, 0.05s elapsed
Nmap scan report for 192.168.56.101
Host is up (0.0012s latency).

PORT    STATE SERVICE
445/tcp open  microsoft-ds

Host script results:
| smb-protocols:
|   dialects:
|     2.02
|     2.10
|     3.00
|     3.02
|_    3.11

Nmap scan report for 192.168.56.103
Host is up (0.00042s latency).

PORT    STATE SERVICE
445/tcp open  microsoft-ds

Host script results:
| smb-protocols:
|   dialects:
|     NT LM 0.12 (SMBv1) [dangerous, but default]
|     2.02
|_    2.10

NSE: Script Post-scanning.
Initiating NSE at 14:30
Completed NSE at 14:30, 0.00s elapsed
Read data files from: /usr/local/bin/../share/nmap
Nmap done: 2 IP addresses (2 hosts up) scanned in 0.27 seconds
wvu@kharak:~$
```

I'm still inclined to land your PR.