@NoodleOfDeath 2019-05-22
Master still does not work on my target, but with the fix it does. I have no idea why because I am currently taking the PWK course and still have not rooted the machine (I'm a noob). Below I include all of my logs:

**Without Fix**
```
msf5 > git status
[*] exec: git status

On branch master
Your branch is up to date with 'origin/master'.

nothing to commit, working tree clean
msf5 > use auxiliary/scanner/smb/pipe_auditor
msf5 auxiliary(scanner/smb/pipe_auditor) > options

Module options (auxiliary/scanner/smb/pipe_auditor):

   Name         Current Setting                                                         Required  Description
   ----         ---------------                                                         --------  -----------
   NAMED_PIPES  /root/OneDrive/git/metasploit-framework/data/wordlists/named_pipes.txt  yes       List of named pipes to check
   RHOSTS                                                                               yes       The target address range or CIDR identifier
   SMBDomain    .                                                                       no        The Windows domain to use for authentication
   SMBPass                                                                              no        The password for the specified username
   SMBUser                                                                              no        The username to authenticate as
   THREADS      1                                                                       yes       The number of concurrent threads

msf5 auxiliary(scanner/smb/pipe_auditor) > cat /root/OneDrive/git/metasploit-framework/data/wordlists/named_pipes.txt
[*] exec: cat /root/OneDrive/git/metasploit-framework/data/wordlists/named_pipes.txt

netlogon
lsarpc
samr
browser
atsvc
DAV RPC SERVICE
epmapper
eventlog
InitShutdown
keysvc
lsass
LSM_API_service
ntsvcs
plugplay
protected_storage
router
SapiServerPipeS-1-5-5-0-70123
scerpc
srvsvc
tapsrv
trkwks
W32TIME_ALT
wkssvc
PIPE_EVENTROOT\CIMV2SCM EVENT PROVIDER
db2remotecmd
msf5 auxiliary(scanner/smb/pipe_auditor) > setg RHOSTS 10.11.1.136
RHOSTS => 10.11.1.136
msf5 auxiliary(scanner/smb/pipe_auditor) > setg VERBOSE true
VERBOSE => true
msf5 auxiliary(scanner/smb/pipe_auditor) > run

[-] 10.11.1.136:139       - Inaccessible named pipe: netlogon - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: lsarpc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: samr - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: browser - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: atsvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: DAV RPC SERVICE - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: epmapper - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: eventlog - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: InitShutdown - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: keysvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: lsass - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: LSM_API_service - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: ntsvcs - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: plugplay - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: protected_storage - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: router - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: SapiServerPipeS-1-5-5-0-70123 - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: scerpc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: srvsvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: tapsrv - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: netlogon - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: lsarpc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: samr - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: browser - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: atsvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: DAV RPC SERVICE - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: epmapper - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: eventlog - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: InitShutdown - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: keysvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: lsass - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: LSM_API_service - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: ntsvcs - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: plugplay - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: protected_storage - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: router - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: SapiServerPipeS-1-5-5-0-70123 - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: scerpc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: srvsvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: tapsrv - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: trkwks - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: W32TIME_ALT - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: wkssvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: PIPE_EVENTROOT\CIMV2SCM EVENT PROVIDER - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:445       - Inaccessible named pipe: db2remotecmd - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 10.11.1.136:          - Scanned 1 of 1 hosts (100% complete)
[*] Auxiliary module execution completed
```

**Nmap Scan of Target**
```
root@KaliNoodle:~/labs/133# nmap -Pn -n -v -p 445 --script smb-protocols 10.11.1.136
Starting Nmap 7.70 ( https://nmap.org ) at 2019-05-22 13:14 EDT
NSE: Loaded 1 scripts for scanning.
NSE: Script Pre-scanning.
Initiating NSE at 13:14
Completed NSE at 13:14, 0.00s elapsed
Initiating ARP Ping Scan at 13:14
Scanning 10.11.1.136 [1 port]
Completed ARP Ping Scan at 13:14, 0.15s elapsed (1 total hosts)
Initiating SYN Stealth Scan at 13:14
Scanning 10.11.1.136 [1 port]
Discovered open port 445/tcp on 10.11.1.136
Completed SYN Stealth Scan at 13:14, 0.27s elapsed (1 total ports)
NSE: Script scanning 10.11.1.136.
Initiating NSE at 13:14
Completed NSE at 13:14, 3.95s elapsed
Nmap scan report for 10.11.1.136
Host is up (0.12s latency).

PORT    STATE SERVICE
445/tcp open  microsoft-ds
MAC Address: 00:50:56:93:37:2B (VMware)

Host script results:
| smb-protocols: 
|   dialects: 
|_    NT LM 0.12 (SMBv1) [dangerous, but default]

NSE: Script Post-scanning.
Initiating NSE at 13:14
Completed NSE at 13:14, 0.00s elapsed
Read data files from: /usr/bin/../share/nmap
Nmap done: 1 IP address (1 host up) scanned in 4.63 seconds
           Raw packets sent: 2 (72B) | Rcvd: 2 (72B)
```

**With Fix**
```
msf5 auxiliary(scanner/smb/pipe_auditor) > git checkout smb/client/pipe_auditor
[*] exec: git checkout smb/client/pipe_auditor

Switched to branch 'smb/client/pipe_auditor'
Your branch is up to date with 'origin/smb/client/pipe_auditor'.
msf5 auxiliary(scanner/smb/pipe_auditor) > 
[2]+  Stopped                 ./msfconsole
root@KaliNoodle:~/git/metasploit-framework# ./msfconsole
[-] ***rting the Metasploit Framework console...\
[-] * WARNING: No database support: No database YAML file
[-] ***
                                                  

MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
MMMMMMMMMMM                MMMMMMMMMM
MMMN$                           vMMMM
MMMNl  MMMMM             MMMMM  JMMMM
MMMNl  MMMMMMMN       NMMMMMMM  JMMMM
MMMNl  MMMMMMMMMNmmmNMMMMMMMMM  JMMMM
MMMNI  MMMMMMMMMMMMMMMMMMMMMMM  jMMMM
MMMNI  MMMMMMMMMMMMMMMMMMMMMMM  jMMMM
MMMNI  MMMMM   MMMMMMM   MMMMM  jMMMM
MMMNI  MMMMM   MMMMMMM   MMMMM  jMMMM
MMMNI  MMMNM   MMMMMMM   MMMMM  jMMMM
MMMNI  WMMMM   MMMMMMM   MMMM#  JMMMM
MMMMR  ?MMNM             MMMMM .dMMMM
MMMMNm `?MMM             MMMM` dMMMMM
MMMMMMN  ?MM             MM?  NMMMMMN
MMMMMMMMNe                 JMMMMMNMMM
MMMMMMMMMMNm,            eMMMMMNMMNMM
MMMMNNMNMMMMMNx        MMMMMMNMMNMMNM
MMMMMMMMNMMNMMMMm+..+MMNMMNMNMMNMMNMM
        https://metasploit.com


       =[ metasploit v5.0.23-dev-3ffd6363f3               ]
+ -- --=[ 1891 exploits - 1065 auxiliary - 329 post       ]
+ -- --=[ 546 payloads - 44 encoders - 10 nops            ]
+ -- --=[ 2 evasion                                       ]

[*] Starting persistent handler(s)...
msf5 > use auxiliary/scanner/smb/pipe_auditor
msf5 auxiliary(scanner/smb/pipe_auditor) > setg RHOSTS 10.11.1.136
RHOSTS => 10.11.1.136
msf5 auxiliary(scanner/smb/pipe_auditor) > setg VERBOSE true
VERBOSE => true
msf5 auxiliary(scanner/smb/pipe_auditor) > run

[*] 10.11.1.136:139       - Connected to named pipe: \netlogon
[*] 10.11.1.136:139       - Connected to named pipe: \lsarpc
[*] 10.11.1.136:139       - Connected to named pipe: \samr
[-] 10.11.1.136:139       - Inaccessible named pipe: \browser - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \atsvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \DAV RPC SERVICE - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \epmapper - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 10.11.1.136:139       - Connected to named pipe: \eventlog
[-] 10.11.1.136:139       - Inaccessible named pipe: \InitShutdown - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \keysvc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 10.11.1.136:139       - Connected to named pipe: \lsass
[-] 10.11.1.136:139       - Inaccessible named pipe: \LSM_API_service - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 10.11.1.136:139       - Connected to named pipe: \ntsvcs
[-] 10.11.1.136:139       - Inaccessible named pipe: \plugplay - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \protected_storage - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \router - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \SapiServerPipeS-1-5-5-0-70123 - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \scerpc - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 10.11.1.136:139       - Connected to named pipe: \srvsvc
[-] 10.11.1.136:139       - Inaccessible named pipe: \tapsrv - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \trkwks - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \W32TIME_ALT - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[*] 10.11.1.136:139       - Connected to named pipe: \wkssvc
[-] 10.11.1.136:139       - Inaccessible named pipe: \PIPE_EVENTROOT\CIMV2SCM EVENT PROVIDER - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[-] 10.11.1.136:139       - Inaccessible named pipe: \db2remotecmd - The server responded with error: STATUS_OBJECT_NAME_NOT_FOUND (Command=162 WordCount=0)
[+] 10.11.1.136:139       - Pipes: \\netlogon, \\lsarpc, \\samr, \\eventlog, \\lsass, \\ntsvcs, \\srvsvc, \\wkssvc
[*] 10.11.1.136:          - Scanned 1 of 1 hosts (100% complete)
[*] Auxiliary module execution completed
msf5 auxiliary(scanner/smb/pipe_auditor) > 
```